+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFileSystemConfigurationTest.java
@@ -14,10 +14,12 @@
 
 package com.google.cloud.hadoop.fs.gcs;
 
+import static com.google.cloud.hadoop.fs.gcs.GoogleHadoopFileSystemConfiguration.GCS_ROOT_URL;
 import static com.google.common.truth.Truth.assertThat;
 import static com.google.common.truth.Truth.assertWithMessage;
 import static org.junit.Assert.assertThrows;
 
+import com.google.api.services.storage.Storage;
 import com.google.cloud.hadoop.fs.gcs.GoogleHadoopFileSystemBase.GcsFileChecksumType;
 import com.google.cloud.hadoop.fs.gcs.GoogleHadoopFileSystemBase.OutputStreamType;
 import com.google.cloud.hadoop.gcsio.GoogleCloudStorageFileSystemOptions;
@@ -51,6 +53,7 @@ public class GoogleHadoopFileSystemConfigurationTest {
           put("fs.gs.copy.with.rewrite.enable", true);
           put("fs.gs.rewrite.max.bytes.per.call", 536870912);
           put("fs.gs.config.override.file", null);
+          put("fs.gs.storage.root.url", Storage.DEFAULT_ROOT_URL);
           put("fs.gs.reported.permissions", "700");
           put("fs.gs.delegation.token.binding", null);
           put("fs.gs.bucket.delete.enable", false);
@@ -208,5 +211,16 @@ public class GoogleHadoopFileSystemConfigurationTest {
       throw new RuntimeException(
           String.format("Failed to get '%s' field value", field.getName()), e);
     }
+  }
+
+  @Test
+  public void customPropertiesValues() {
+    Configuration config = new Configuration();
+    config.set(GCS_ROOT_URL.getKey(), "https://unit-test-storage.googleapis.com/");
+
+    GoogleCloudStorageOptions options =
+        GoogleHadoopFileSystemConfiguration.getGcsOptionsBuilder(config).build();
+
+    assertThat(options.getStorageRootUrl()).isEqualTo("https://unit-test-storage.googleapis.com/");
   }
 }