+++ util/src/main/java/com/google/cloud/hadoop/util/CredentialOptions.java
@@ -16,6 +16,7 @@ package com.google.cloud.hadoop.util;
 
 import static com.google.cloud.hadoop.util.HttpTransportFactory.toSecretString;
 
+import com.google.api.client.googleapis.auth.oauth2.GoogleOAuthConstants;
 import com.google.auto.value.AutoValue;
 import com.google.cloud.hadoop.util.HttpTransportFactory.HttpTransportType;
 import javax.annotation.Nullable;
@@ -23,6 +24,7 @@ import javax.annotation.Nullable;
 /** Configuration for how components should obtain Credentials. */
 @AutoValue
 public abstract class CredentialOptions {
+
   static final boolean SERVICE_ACCOUNT_ENABLED_DEFAULT = true;
 
   static final boolean NULL_CREDENTIALS_ENABLED_DEFAULT = false;
@@ -30,11 +32,14 @@ public abstract class CredentialOptions {
   static final HttpTransportType HTTP_TRANSPORT_TYPE_DEFAULT =
       HttpTransportFactory.DEFAULT_TRANSPORT_TYPE;
 
+  static final String TOKEN_SERVER_URL_DEFAULT = GoogleOAuthConstants.TOKEN_SERVER_URL;
+
   public static Builder builder() {
     return new AutoValue_CredentialOptions.Builder()
         .setServiceAccountEnabled(SERVICE_ACCOUNT_ENABLED_DEFAULT)
         .setNullCredentialEnabled(NULL_CREDENTIALS_ENABLED_DEFAULT)
-        .setTransportType(HTTP_TRANSPORT_TYPE_DEFAULT);
+        .setTransportType(HTTP_TRANSPORT_TYPE_DEFAULT)
+        .setTokenServerUrl(TOKEN_SERVER_URL_DEFAULT);
   }
 
   public abstract boolean isServiceAccountEnabled();
@@ -75,6 +80,8 @@ public abstract class CredentialOptions {
 
   public abstract HttpTransportType getTransportType();
 
+  public abstract String getTokenServerUrl();
+
   @Nullable
   public abstract String getProxyAddress();
 
@@ -100,6 +107,7 @@ public abstract class CredentialOptions {
         + ("oAuthCredentialFile: " + getOAuthCredentialFile() + '\n')
         + ("nullCredentialEnabled: " + isNullCredentialEnabled() + '\n')
         + ("transportType: " + getTransportType() + '\n')
+        + ("tokenServerUrl: " + getTokenServerUrl() + '\n')
         + ("proxyAddress: " + getProxyAddress() + '\n')
         + ("proxyUsername: " + toSecretString(getProxyUsername()) + '\n')
         + ("proxyPassword: " + toSecretString(getProxyPassword()) + '\n')
@@ -130,6 +138,8 @@ public abstract class CredentialOptions {
     public abstract Builder setNullCredentialEnabled(boolean nullCredentialEnabled);
 
     public abstract Builder setTransportType(HttpTransportType transportType);
+
+    public abstract Builder setTokenServerUrl(String tokenServerUrl);
 
     public abstract Builder setProxyAddress(String proxyAddress);