+++ bigquery/CHANGES.md
@@ -8,6 +8,11 @@
 
 1.  Remove Log4j backend for Google Flogger.
 
+1.  Add properties to override Google Cloud API endpoints: 
+
+        mapred.bq.bigquery.root.url
+        mapred.bq.token.server.url
+
 ### 1.0.1 - 2020-02-13
 
 1.  Fix shaded jar - add back missing relocated dependencies.